made errors its own module
